Include missing lxml dep
Since PyCortexMDebug is now included as a submodule, need this
dependency in the top-level list.

Restrict python versions a little to enable installing a pre-built lxml
wheel, which is loads faster than waiting for a source dist to build
(and removes the need for build deps on the target).

Bump version to 0.0.5.